OS X binary release
Please remove PingJiao.app.zip from this repo. It is an executable for
OS X.
